// src/interfaces/userRoles.ts
export interface IPermissonList {
  Name?: string;
  Description?: string;
  PosBilling: boolean;
  OrderManagement: boolean;
  Quotation: boolean;
  Journal: boolean;
  Purchase: boolean;
  Invoice: boolean;
  Receipt: boolean;
  Payment: boolean;
  BankCash: boolean;
  SalesReturn: boolean;
  ImportBill: boolean;
  PurchaseReturn: boolean;
  PurchaseOrder: boolean;
  Ledger: boolean;
  GroupLedger: boolean;
  Products: boolean;
  Category: boolean;
  UnitType: boolean;
  TrialBalance: boolean;
  ProfitLoss: boolean;
  BalanceSheet: boolean;
  SalesDateWise: boolean;
  ItemWiseProfit: boolean;
  SalesBook: boolean;
  MaterializedView: boolean;
  SalesReturnBook: boolean;
  ChangePassword: boolean;
  Users: boolean;
  Roles: boolean;
  Departments: boolean;
  FinancialYear: boolean;
  Company: boolean;
  Warehouse: boolean;
  WarehouseType: boolean;
  Branch: boolean;
  AssignRoles: boolean;
  TransactionType: boolean;
  MasterLedger: boolean;
  MasterLedger2: boolean;
  PurchaseBook: boolean;
  ItemStockLedger: boolean;
  StockInHand: boolean;
  //added new
  InvAdd: boolean;
  InvEdit: boolean;
  InvDelete: boolean;
  InvView: boolean;
  PurAdd: boolean;
  PurEdit: boolean;
  PurDelete: boolean;
  PurView: boolean;
  ReceiptAdd: boolean;
  ReceiptEdit: boolean;
  ReceiptDelete: boolean;
  ReceiptView: boolean;
  PaymentAdd: boolean;
  PaymentEdit: boolean;
  PaymentDelete: boolean;
  PaymentView: boolean;
  BankAdd: boolean;
  BankEdit: boolean;
  BankDelete: boolean;
  BankView: boolean;
  SRAdd: boolean;
  SREdit: boolean;
  SRDelete: boolean;
  SRView: boolean;
  PRAdd: boolean;
  PREdit: boolean;
  PRDelete: boolean;
  PRView: boolean;
  JournalAdd: boolean;
  JournalEdit: boolean;
  JournalDelete: boolean;
  JournalView: boolean;
  OMAdd: boolean;
  OMEdit: boolean;
  OMDelete: boolean;
  OMView: boolean;
  OMInvoice: boolean;
  QuotationAdd: boolean;
  QuotationEdit: boolean;
  QuotationDelete: boolean;
  QuotationView: boolean;
  QuotationInvoice: boolean;
  POAdd: boolean;
  POEdit: boolean;
  PODelete: boolean;
  POView: boolean;
  MLAdd: boolean;
  ProductAdd: boolean;
  CategoryAdd: boolean;
  UTAdd: boolean;
  WHTAdd: boolean;
  WHAdd: boolean;
  WHEdit: boolean;
  DeAdd: boolean;
  DeEdit: boolean;
  FYAdd: boolean;
  BranchAdd: boolean;
  UserAdd: boolean;
  RoleAdd: boolean;
  GetBranch: boolean;
  GetDepartment: boolean;
  GetWarehouse: boolean;
  GetCustomers: boolean;
  GetProducts: boolean;
  GetMasterLedger: boolean;
  GetSourceAccount: boolean;
  GetRefNo: boolean;
  SMS: boolean;
  Audit: boolean;
}
export interface IUserRole {
  Users: User[];
  RoleName?: any;
  Description: string;
  PermissionList: string;
  Selected: boolean;
  IsAdd?: any;
  IsEdit?: any;
  IsDelete?: any;
  IsView?: any;
  CreatedOn?: any;
  CreatedBy?: any;
  LastChangedDate?: any;
  LastChangedBy?: any;
  IsSysAdmin: boolean;
  Id: number;
  Name: string;
}

interface User {
  UserId: string;
  RoleId: string;
}